+++ pumpkin/src/plugin/api/events/block/mod.rs
@@ -2,7 +2,7 @@ pub mod block_break;
 pub mod block_burn;
 pub mod block_can_build;
 pub mod block_place;
-pub mod block_redstone;
+pub mod redstone;
 
 use pumpkin_data::Block;
